update transaction functions
