+++ src/features/transaction/TransactionTableTools.tsx
@@ -1,7 +1,6 @@
 import { Button } from "@chakra-ui/react";
 import { useDeleteTransaction } from "../../api/endpoints/transaction/transactions.api.ts";
 import i18next from "../../tools/language/language.ts";
-import {useGetUserTransactionInfo} from "../../api/endpoints/user/transaction/userTransaction.api.ts";
 import useTransactionEditDrawer from "../../stores/drawerStore.ts";
 
 type Props = {
